+++ src/main/java/blackjack/model/CardShuffler.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 import java.util.List;
