+++ src/main/java/blackjack/model/Deck.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 import java.util.LinkedList;
 import java.util.List;
