+++ src/main/java/blackjack/model/Suit.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 public enum Suit {
