+++ src/test/java/blackjack/model/GameTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
